email_notification/requirements.txt
google-api-python-client==1.7.8
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
google-cloud-secret-manager
google-cloud-storage
protobuf==3.20.0